+++ src/devtools/index.ts
@@ -1,0 +1,5 @@
+export {
+  FormularDevtools,
+  type FormularDevtoolsProps,
+  type FormularDevtoolsPosition,
+} from './FormularDevtools'